+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Nexus2](https://nexus.inventage.com/#nexus-search;gav~~~???~~) - [Nexus3](https://nexus3.inventage.com/#browse/search=version%3D???)
 
 ### Added
+
 - Add custom claim check for prefix urls `/ips/portal-database`and `/ips/portal-messaging`. Only jwt with valid claims can access those urls.
 ([PORTAL-626](https://issue.inventage.com/browse/PORTAL-626))
 - Extend the claim verification in the bearer only middleware. It should be possible to verify arbitrary claims.
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - OpenTelemetry: traces and spans are created using opentelemetry sdk and extension of vertx. The configuration of the opentelemetry logic occurs using [opentelemetry sdk autoconfiguration](https://github.com/open-telemetry/opentelemetry-java/tree/main/sdk-extensions/autoconfigure) ([PORTAL-642](https://issue.inventage.com/browse/PORTAL-642)).
 - HTTP header `X-Forwarded-Port´ is set for outgoing requests to services ([PORTAL-713](https://issue.inventage.com/browse/PORTAL-713)).
 - Add route to portal-monitoring UI ([PORTAL-695](https://issue.inventage.com/browse/PORTAL-695))
+- Allow literal dot `.` characters in `Path` and `PathPrefix` and `Host` rule values in the router factory.
 
 ### Removed